topic count by bytes / sizeof(int32_t) not sizeof(long).
this is why if you typecast bytes to struct you use stdint.h not assume
sizeof(long) == 4.

Pulled from:
https://github.com/joncampbell123/helpdeco/commit/65dfecb19ddc9664117b1b418db9981fbc148731

+++ helpdeco.c
@@ -3851,7 +3851,7 @@ void ContextLoad(FILE *HelpFile)
     {
 	Topic=my_malloc(entries);
 	my_fread(Topic,entries,HelpFile);
-	Topics=(int)(entries/sizeof(long));
+	Topics=(int)(entries/sizeof(int32_t));
     }
 }
